Code cleanup

+++ deploy.sh
@@ -1,2 +1,5 @@
-rm code.zip
+if [ -f code.zip ]; then
+    rm code.zip
+fi
+
 zip -r code.zip ./*.png ./*.js ./*.html ./*.json
